Implemented HTTPS wrapper issue workaround to add GitHub events feed
As per issue 1, I have implemented using a HTTP proxy server to do the
HTTPS request on behalf of the web server to return a GitHub events
feed. The PHP code displays the results as a bullet point list and the
CSS makes it fit into a fixed height DIV with a vertical scrollbar.